Move downloader specific types to the downloader
