+++ src/sources.rs
@@ -1,6 +1,6 @@
 //! Download sources types and related implementations.
 
-use anyhow::{Error, Result};
+use anyhow::Result;
 use std::collections::BTreeMap;
 use std::fmt;
 use std::path::PathBuf;
@@ -9,7 +9,7 @@ use std::sync::Arc;
 use serde::{Deserialize, Deserializer, Serialize};
 use url::Url;
 
-use crate::types::{Glob, ObjectFileSourceURI, ObjectId, ObjectType};
+use crate::types::{Glob, ObjectId, ObjectType};
 use crate::utils::paths;
 use crate::utils::sentry::WriteSentryScope;
 
@@ -100,40 +100,6 @@ impl WriteSentryScope for SourceConfig {
     }
 }
 
-/// A location for a file retrievable from many source configs.
-///
-/// It is essentially a `/`-separated string. This is currently used by all sources other than
-/// [`SentrySourceConfig`]. This may change in the future.
-#[derive(Debug, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
-pub struct SourceLocation(String);
-
-impl SourceLocation {
-    pub fn new(loc: impl Into<String>) -> Self {
-        SourceLocation(loc.into())
-    }
-
-    /// Return an iterator of the location segments.
-    pub fn segments<'a>(&'a self) -> impl Iterator<Item = &str> + 'a {
-        self.0.split('/').filter(|s| !s.is_empty())
-    }
-}
-
-impl fmt::Display for SourceLocation {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-/// An identifier for a file retrievable from a [`SentrySourceConfig`].
-#[derive(Debug, Clone, Eq, PartialEq, Hash, Deserialize)]
-pub struct SentryFileId(String);
-
-impl fmt::Display for SentryFileId {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
 /// Configuration for the Sentry-internal debug files endpoint.
 #[derive(Deserialize, Clone, Debug)]
 pub struct SentrySourceConfig {
@@ -177,311 +143,6 @@ pub struct FilesystemSourceConfig {
 
     #[serde(flatten)]
     pub files: CommonSourceConfig,
-}
-
-/// Represents a single object file stored on a source.
-///
-/// This joins the file location together with a [`SourceConfig`] and thus provides all
-/// information to retrieve the object file from its source.
-#[derive(Debug, Clone)]
-pub enum ObjectFileSource {
-    Sentry(SentryObjectFileSource),
-    Http(HttpObjectFileSource),
-    S3(S3ObjectFileSource),
-    Gcs(GcsObjectFileSource),
-    Filesystem(FilesystemObjectFileSource),
-}
-
-impl fmt::Display for ObjectFileSource {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            ObjectFileSource::Sentry(ref s) => {
-                write!(f, "Sentry source '{}' file id '{}'", s.source.id, s.file_id)
-            }
-            ObjectFileSource::Http(ref s) => {
-                write!(f, "HTTP source '{}' location '{}'", s.source.id, s.location)
-            }
-            ObjectFileSource::S3(ref s) => {
-                write!(f, "S3 source '{}' location '{}'", s.source.id, s.location)
-            }
-            ObjectFileSource::Gcs(ref s) => {
-                write!(f, "GCS source '{}' location '{}'", s.source.id, s.location)
-            }
-            ObjectFileSource::Filesystem(ref s) => {
-                write!(
-                    f,
-                    "Filesystem source '{}' location '{}'",
-                    s.source.id, s.location
-                )
-            }
-        }
-    }
-}
-
-impl ObjectFileSource {
-    /// Whether debug files from this source may be shared.
-    pub fn is_public(&self) -> bool {
-        match self {
-            ObjectFileSource::Sentry(_) => false,
-            ObjectFileSource::Http(ref x) => x.source.files.is_public,
-            ObjectFileSource::S3(ref x) => x.source.files.is_public,
-            ObjectFileSource::Gcs(ref x) => x.source.files.is_public,
-            ObjectFileSource::Filesystem(ref x) => x.source.files.is_public,
-        }
-    }
-
-    pub fn cache_key(&self) -> String {
-        match self {
-            ObjectFileSource::Sentry(ref x) => {
-                format!("{}.{}.sentryinternal", x.source.id, x.file_id)
-            }
-            ObjectFileSource::Http(ref x) => {
-                format!("{}.{}", x.source.id, x.location)
-            }
-            ObjectFileSource::S3(ref x) => {
-                format!("{}.{}", x.source.id, x.location)
-            }
-            ObjectFileSource::Gcs(ref x) => {
-                format!("{}.{}", x.source.id, x.location)
-            }
-            ObjectFileSource::Filesystem(ref x) => {
-                format!("{}.{}", x.source.id, x.location)
-            }
-        }
-    }
-
-    /// Returns the ID of the source.
-    ///
-    /// Within one request these IDs should be unique, this includes any sources from the
-    /// configuration which are available to all requests.
-    pub fn source_id(&self) -> &SourceId {
-        match self {
-            ObjectFileSource::Sentry(ref x) => &x.source.id,
-            ObjectFileSource::Http(ref x) => &x.source.id,
-            ObjectFileSource::S3(ref x) => &x.source.id,
-            ObjectFileSource::Gcs(ref x) => &x.source.id,
-            ObjectFileSource::Filesystem(ref x) => &x.source.id,
-        }
-    }
-
-    pub fn source_type_name(&self) -> &'static str {
-        match *self {
-            ObjectFileSource::Sentry(..) => "sentry",
-            ObjectFileSource::S3(..) => "s3",
-            ObjectFileSource::Gcs(..) => "gcs",
-            ObjectFileSource::Http(..) => "http",
-            ObjectFileSource::Filesystem(..) => "filesystem",
-        }
-    }
-
-    /// Returns a URI for the location of the object file.
-    ///
-    /// There is no guarantee about any format of this URI, for some sources it could be
-    /// very abstract.  In general the source should try and producde a URI which can be
-    /// used directly into the source-specific tooling.  E.g. for an HTTP source this would
-    /// be an `http://` or `https://` URL, for AWS S3 it would be an `s3://` url etc.
-    pub fn uri(&self) -> ObjectFileSourceURI {
-        match self {
-            ObjectFileSource::Sentry(ref file_source) => file_source.uri(),
-            ObjectFileSource::Http(ref file_source) => file_source.uri(),
-            ObjectFileSource::S3(ref file_source) => file_source.uri(),
-            ObjectFileSource::Gcs(ref file_source) => file_source.uri(),
-            ObjectFileSource::Filesystem(ref file_source) => file_source.uri(),
-        }
-    }
-}
-
-impl WriteSentryScope for ObjectFileSource {
-    fn write_sentry_scope(&self, scope: &mut ::sentry::Scope) {
-        scope.set_tag("source.id", self.source_id());
-        scope.set_tag("source.type", self.source_type_name());
-        scope.set_tag("source.is_public", self.is_public());
-    }
-}
-
-impl From<SentryObjectFileSource> for ObjectFileSource {
-    fn from(source: SentryObjectFileSource) -> Self {
-        Self::Sentry(source)
-    }
-}
-
-impl From<HttpObjectFileSource> for ObjectFileSource {
-    fn from(source: HttpObjectFileSource) -> Self {
-        Self::Http(source)
-    }
-}
-
-impl From<S3ObjectFileSource> for ObjectFileSource {
-    fn from(source: S3ObjectFileSource) -> Self {
-        Self::S3(source)
-    }
-}
-
-impl From<GcsObjectFileSource> for ObjectFileSource {
-    fn from(source: GcsObjectFileSource) -> Self {
-        Self::Gcs(source)
-    }
-}
-
-impl From<FilesystemObjectFileSource> for ObjectFileSource {
-    fn from(source: FilesystemObjectFileSource) -> Self {
-        Self::Filesystem(source)
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct SentryObjectFileSource {
-    pub source: Arc<SentrySourceConfig>,
-    pub file_id: SentryFileId,
-}
-
-impl SentryObjectFileSource {
-    pub fn new(source: Arc<SentrySourceConfig>, file_id: SentryFileId) -> Self {
-        Self { source, file_id }
-    }
-
-    pub fn uri(&self) -> ObjectFileSourceURI {
-        self.url().as_str().into()
-    }
-
-    /// Returns the URL from which to download this object file.
-    pub fn url(&self) -> Url {
-        let mut url = self.source.url.clone();
-        url.query_pairs_mut().append_pair("id", &self.file_id.0);
-        url
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct HttpObjectFileSource {
-    pub source: Arc<HttpSourceConfig>,
-    pub location: SourceLocation,
-}
-
-impl HttpObjectFileSource {
-    pub fn new(source: Arc<HttpSourceConfig>, location: SourceLocation) -> Self {
-        Self { source, location }
-    }
-
-    pub fn uri(&self) -> ObjectFileSourceURI {
-        match self.url() {
-            Ok(url) => url.as_ref().into(),
-            Err(_) => "".into(),
-        }
-    }
-
-    /// Returns the URL from which to download this object file.
-    pub fn url(&self) -> Result<Url> {
-        join_url_encoded(&self.source.url, &self.location)
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct S3ObjectFileSource {
-    pub source: Arc<S3SourceConfig>,
-    pub location: SourceLocation,
-}
-
-impl S3ObjectFileSource {
-    pub fn new(source: Arc<S3SourceConfig>, location: SourceLocation) -> Self {
-        Self { source, location }
-    }
-
-    /// Returns the S3 key.
-    ///
-    /// This is equivalent to the pathname within the bucket.
-    pub fn key(&self) -> String {
-        join_prefix_location(&self.source.prefix, &self.location)
-    }
-
-    /// Returns the S3 bucket name.
-    pub fn bucket(&self) -> String {
-        self.source.bucket.clone()
-    }
-
-    /// Returns the `s3://` URI from which to download this object file.
-    pub fn uri(&self) -> ObjectFileSourceURI {
-        format!(
-            "s3://{bucket}/{key}",
-            bucket = &self.source.bucket,
-            key = self.key()
-        )
-        .into()
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct GcsObjectFileSource {
-    pub source: Arc<GcsSourceConfig>,
-    pub location: SourceLocation,
-}
-
-impl GcsObjectFileSource {
-    pub fn new(source: Arc<GcsSourceConfig>, location: SourceLocation) -> Self {
-        Self { source, location }
-    }
-
-    /// Returns the S3 key.
-    ///
-    /// This is equivalent to the pathname within the bucket.
-    pub fn key(&self) -> String {
-        join_prefix_location(&self.source.prefix, &self.location)
-    }
-
-    /// Returns the `gs://` URI from which to download this object file.
-    pub fn uri(&self) -> ObjectFileSourceURI {
-        format!(
-            "gs://{bucket}/{key}",
-            bucket = &self.source.bucket,
-            key = self.key()
-        )
-        .into()
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct FilesystemObjectFileSource {
-    pub source: Arc<FilesystemSourceConfig>,
-    pub location: SourceLocation,
-}
-
-impl FilesystemObjectFileSource {
-    pub fn new(source: Arc<FilesystemSourceConfig>, location: SourceLocation) -> Self {
-        Self { source, location }
-    }
-
-    /// Returns the path from which to fetch this object file.
-    pub fn path(&self) -> PathBuf {
-        self.source.path.join(&self.location.0)
-    }
-
-    /// Returns the `file://` URI from which to fetch this object file.
-    ///
-    /// This is a quick-and-dirty approximation, not fully RFC8089-compliant.  E.g. we do
-    /// not provide a hostname nor percent-encode.  Use this only for diagnostics and use
-    /// [`FilesystemObjectFileSource::path`] if the actual file location is needed.
-    pub fn uri(&self) -> ObjectFileSourceURI {
-        format!("file:///{path}", path = self.path().display()).into()
-    }
-}
-
-/// Joins the relative path to the given URL.
-///
-/// As opposed to [`Url::join`], this only supports relative paths. Each segment of the path is
-/// percent-encoded. Empty segments are skipped, for example, `foo//bar` is collapsed to `foo/bar`.
-///
-/// The base URL is treated as directory. If it does not end with a slash, then a slash is
-/// automatically appended.
-///
-/// Returns `Err` if the URL is cannot-be-a-base.
-fn join_url_encoded(base: &Url, path: &SourceLocation) -> Result<Url> {
-    let mut joined = base.clone();
-    joined
-        .path_segments_mut()
-        .map_err(|_| Error::msg("URL cannot-be-a-base"))?
-        .pop_if_empty()
-        .extend(path.segments());
-    Ok(joined)
 }
 
 /// Local helper to deserializes an S3 region string in `S3SourceKey`.
@@ -559,15 +220,6 @@ pub struct GcsSourceConfig {
 
     #[serde(flatten)]
     pub files: CommonSourceConfig,
-}
-
-fn join_prefix_location(prefix: &str, location: &SourceLocation) -> String {
-    let trimmed = prefix.trim_matches(&['/'][..]);
-    if trimmed.is_empty() {
-        location.0.clone()
-    } else {
-        format!("{}/{}", trimmed, location.0)
-    }
 }
 
 /// Configuration for S3 symbol buckets.
@@ -785,90 +437,5 @@ impl AsRef<str> for FileType {
             FileType::Breakpad => "breakpad",
             FileType::SourceBundle => "sourcebundle",
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_join_prefix_location() {
-        let key = join_prefix_location(&String::from(""), &SourceLocation::new("spam/ham"));
-        assert_eq!(key, "spam/ham");
-
-        let key = join_prefix_location(&String::from("eggs"), &SourceLocation::new("spam/ham"));
-        assert_eq!(key, "eggs/spam/ham");
-
-        let key = join_prefix_location(&String::from("/eggs/bacon/"), &SourceLocation::new("spam"));
-        assert_eq!(key, "eggs/bacon/spam");
-
-        let key = join_prefix_location(&String::from("//eggs//"), &SourceLocation::new("spam"));
-        assert_eq!(key, "eggs/spam");
-    }
-
-    #[test]
-    fn test_sentry_source_download_url() {
-        let source = SentrySourceConfig {
-            id: SourceId::new("test"),
-            url: Url::parse("https://example.net/endpoint/").unwrap(),
-            token: "token".into(),
-        };
-        let file_source =
-            SentryObjectFileSource::new(Arc::new(source), SentryFileId("abc123".into()));
-        let url = file_source.url();
-        assert_eq!(url.as_str(), "https://example.net/endpoint/?id=abc123");
-    }
-
-    #[test]
-    fn test_join_empty() {
-        let base = Url::parse("https://example.org/base").unwrap();
-        let joined = join_url_encoded(&base, &SourceLocation::new("")).unwrap();
-        assert_eq!(joined, "https://example.org/base".parse().unwrap());
-    }
-
-    #[test]
-    fn test_join_space() {
-        let base = Url::parse("https://example.org/base").unwrap();
-        let joined = join_url_encoded(&base, &SourceLocation::new("foo bar")).unwrap();
-        assert_eq!(
-            joined,
-            "https://example.org/base/foo%20bar".parse().unwrap()
-        );
-    }
-
-    #[test]
-    fn test_join_multiple() {
-        let base = Url::parse("https://example.org/base").unwrap();
-        let joined = join_url_encoded(&base, &SourceLocation::new("foo/bar")).unwrap();
-        assert_eq!(joined, "https://example.org/base/foo/bar".parse().unwrap());
-    }
-
-    #[test]
-    fn test_join_trailing_slash() {
-        let base = Url::parse("https://example.org/base/").unwrap();
-        let joined = join_url_encoded(&base, &SourceLocation::new("foo")).unwrap();
-        assert_eq!(joined, "https://example.org/base/foo".parse().unwrap());
-    }
-
-    #[test]
-    fn test_join_leading_slash() {
-        let base = Url::parse("https://example.org/base").unwrap();
-        let joined = join_url_encoded(&base, &SourceLocation::new("/foo")).unwrap();
-        assert_eq!(joined, "https://example.org/base/foo".parse().unwrap());
-    }
-
-    #[test]
-    fn test_join_multi_slash() {
-        let base = Url::parse("https://example.org/base").unwrap();
-        let joined = join_url_encoded(&base, &SourceLocation::new("foo//bar")).unwrap();
-        assert_eq!(joined, "https://example.org/base/foo/bar".parse().unwrap());
-    }
-
-    #[test]
-    fn test_join_absolute() {
-        let base = Url::parse("https://example.org/").unwrap();
-        let joined = join_url_encoded(&base, &SourceLocation::new("foo")).unwrap();
-        assert_eq!(joined, "https://example.org/foo".parse().unwrap());
     }
 }
